backend/requirements.txt
fastapi
uvicorn
openai-whisper
transformers
sentencepiece
sacremoses
pydub
ffmpeg-python
python-multipart
python-dotenv
gTTS
googletrans==4.0.0rc1
torch
playsound==1.2.2
aiofiles